kill all processes in /etc/inittab BEFORE umounting
This is done in /etc/inittab and not in a separate script because the
process which does the "kill -1" needs to have the PID 1, otherwise it
would die trying.

Of course the proper way to do it would be to modify busybox init to
allow for executing code *after* it killed all processes on shutdown,
not just *before* it kills them.  If "umount -ar" is run before
everything is killed, it can fail because programs may still access
files in those mount points.